/dev/sdXn => /dev/sdX for the system images
It seems that the system expects the first partition on the boot medium
(e.g. /dev/sdb1) to contain the OS image and so on. In our case, it must be
the device itself instead (e.g. /dev/sdb).

user@pay2:/mnt/dev/rosa-pkgs/dracut$ git log dracut-041-fixup-boot-device.patch
commit 4dd5adf4fa925410477345188f639b421b0d744c
Author: Eugene Shatokhin <eugene.shatokhin@rosalab.ru>
Date:   Thu Jun 18 11:21:45 2015 +0300

    Fixed the patch: the device may be passed not as /dev/sdXn

    Looks like /dev/disk/by-label/... is currently used, so we need to
    canonicalize the path first.

commit a14a17f992fc295f7e2706ebdae44d2bffc2eb43
Author: Eugene Shatokhin <eugene.shatokhin@rosalab.ru>
Date:   Wed Jun 17 18:10:32 2015 +0300

    Restored and re-diffed a patch to fix boot in Live mode on x32

    Details: http://bugs.rosalinux.ru/show_bug.cgi?id=5768#c3

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>