network: add option to specify preferred network module
Most desktop Linux systems have NetworkManager, many of them have systemd-networkd
inside the systemd package and many of them have dhclient installed.

Current code oft he network meta module, if no specific network module is included,
pulls NetworkManager and its dependency dbus into the initramfs.

I want to avoid such complex systems inside initrd and want to avoid necessity to rebuild
initrds when NetworkManager and dbus packages are updated to keep them in the initrd and
on the host in sync, and also want to standartize initrds by having the same network
module in most of them.

If I add a specific network module into the distribution (ROSA) default config of dracut,
e.g. let it be systemd-networkd, dracut will try to include it even on systems where
systemd-networkd is not installed and fill fail with error.

So adding a config option which allows to change the _preferred_ default if no
specific network module is configured to be added into the initrd.

For now I will set it in ROSA to network-legacy to avoid pulling dbus, maybe it will
have to be changed to systemd-networkd later when this module matures in dracut
or when network-legacy is dropped.

Fixes: https://github.com/dracutdevs/dracut/issues/1404